+++ packages/material-engine/src/index.ts
@@ -1,0 +1,9 @@
+import * as block from './block';
+import * as component from './component';
+import * as material from './material';
+import * as router from './router';
+import * as page from './page';
+import * as scaffold from './scaffold';
+import * as menu from './menu';
+
+export { block, component, material, router, page, scaffold, menu };