+++ extensions/iceworks-material-helper/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.4.1
+
+- fix: ui build error
+
+## 0.4.0
+
+- feat: migrate component-creator, page-creator, page-generator 
+
 ## 0.3.11
 
 - docs: more